requirements.txt
langchain==0.3.26
langchain-community==0.3.26
torch==2.1.1+cpu
langchain-huggingface==0.3.0
openai==1.88.0
chromadb==1.0.13
sentence-transformers==3.2.0
tiktoken==0.9.0

fastapi==0.115.13
uvicorn[standard]==0.34.3

aiogram==3.6.0

beautifulsoup4==4.12.3

pydantic==2.7.4
python-dotenv==1.1.0

gunicorn==22.0.0  

